#400.00: Added BlockChain class and JEST test for class
The BlockChain class creates a chain by adding a genesis block first, and with the addToBlockChain() method, gets the current blockchain, gets the last block, and uses its hash info to create a new block. the new block will contain the new data to store. the JEST test confirms creation of genesis blocks and the ability to add a new block.